pyproject.toml
[project]
name = "git_tool"
authors = [{ name = "Tabea Röthemeyer", email = "tabea.roethemeyer@rub.de" }]
license = { file = "LICENSE" }
description = "Support feature-oriented development workflows with git"
readme = "README.md"
version = "1.0.0"
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Development Status :: 3 - Alpha",
]
dependencies = [
  "typer>=0.12,<0.13",
  "pydantic[email]>=2.8,<3.0",
  "GitPython>=3.1,<4.0",
  "python-dotenv>=1.0,<2.0",
  "prompt_toolkit>=3.0,<4.0",
]

[project.scripts]
git-feature-status = "git_tool.ci.subcommands.feature_status:app"
git-feature-info = "git_tool.ci.subcommands.feature_info:app"
git-feature-add = "git_tool.ci.subcommands.feature_add:app"
git-feature-blame = "git_tool.ci.subcommands.feature_blame:app"
git-feature-commit = "git_tool.ci.subcommands.feature_commit:app"
git-feature-commit-msg = "git_tool.ci.subcommands.feature_commit_msg:app"
git-feature-pre-commit = "git_tool.ci.subcommands.feature_pre_commit:app"
git-feature-commits = "git_tool.ci.subcommands.feature_commits:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["git_tool"]

[tool.pytest.ini_options]
testpaths = ["test"]
addopts = "-s"
pythonpath = ["."]

[tool.black]
line-length = 80
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  tests/data/
  | profiling/
  | scripts/generate_schema.py  # Uses match syntax
)
'''
